rename crate, expose set_mask_for_current
