+++ ccp/Cargo.toml
@@ -15,7 +15,7 @@ ccp-config = { workspace = true }
 ccp-shared = { workspace = true }
 ccp-utils = { workspace = true }
 
-cpu-topology = { workspace = true }
+cpu-utils = { workspace = true }
 randomx-rust-wrapper = { workspace = true }
 
 tokio = { workspace = true }
