+++ ccp/src/cu/proving_thread/sync.rs
@@ -143,7 +143,7 @@ impl ProvingThreadSync {
 
             AsyncToSyncMessage::PinThread(params) => {
                 // TODO: propagate error
-                cpu_topology::pin_current_thread_to(params.core_id);
+                cpu_utils::pinning::pin_current_thread_to(params.core_id);
                 Ok(ThreadState::WaitForMessage)
             }
