+++ ccp/src/cu/proving_thread_utils/round_robin_distributor.rs
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-use cpu_topology::LogicalCoreId;
+use cpu_utils::LogicalCoreId;
 
 use super::threads_distribution_policy::ThreadDistributionPolicy;
