--- crates/cpu-topology/src/lib.rs
@@ -1,40 +1,0 @@
-/*
- * Copyright 2024 Fluence Labs Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#![warn(rust_2018_idioms)]
-#![warn(rust_2021_compatibility)]
-#![deny(
-    dead_code,
-    nonstandard_style,
-    unused_imports,
-    unused_mut,
-    unused_variables,
-    unused_unsafe,
-    unreachable_patterns
-)]
-
-mod cpu_topology;
-mod errors;
-mod thread_pinning;
-mod types;
-
-pub type CTResult<T> = Result<T, CPUTopologyError>;
-
-pub use cpu_topology::CPUTopology;
-pub use errors::CPUTopologyError;
-pub use thread_pinning::pin_current_thread_to;
-pub use types::LogicalCoreId;
-pub use types::PhysicalCoreId;
