+++ crates/cpu-utils/src/errors.rs
