+++ crates/cpu-utils/src/types.rs
